See also https://github.com/NVIDIA-Merlin/HierarchicalKV/pull/26 for past discussions.