@rhdong 2022-11-17
Could we have any benchmark data for evaluating the benefit of this optimization?