@bashimao 2022-12-08
**.:: With change ::.**
```
## Benchmark

* GPU: 1 x NVIDIA A10: 8.6
* Key Type = uint64_t
* Value Type = float32 * {dim}
* Key-Values per OP = 1048576
* ***Throughput Unit: Billion-KV/second***

### On pure HBM mode:

| dim |    capacity | load_factor | HBM(GB) | HMEM(GB) | insert |   find |  erase |
|----:|------------:|------------:|--------:|---------:|-------:|-------:|-------:|
|  16 |    67108864 |        0.50 |      16 |        0 |  0.416 |  0.636 |  0.368 |
|  16 |    67108864 |        0.75 |      16 |        0 |  0.374 |  0.560 |  0.370 |
|  16 |    67108864 |        1.00 |      16 |        0 |  0.102 |  0.334 |  0.321 |
|  64 |    67108864 |        0.50 |      16 |        0 |  0.327 |  0.294 |  0.324 |
|  64 |    67108864 |        0.75 |      16 |        0 |  0.298 |  0.289 |  0.297 |
|  64 |    67108864 |        1.00 |      16 |        0 |  0.097 |  0.254 |  0.341 |

### On HBM+HMEM hybrid mode:

| dim |    capacity | load_factor | HBM(GB) | HMEM(GB) | insert |   find |  erase |
|----:|------------:|------------:|--------:|---------:|-------:|-------:|-------:|
|  64 |   134217728 |        0.50 |      16 |       16 |  0.104 |  0.115 |  0.323 |
|  64 |   134217728 |        0.75 |      16 |       16 |  0.098 |  0.115 |  0.297 |
|  64 |   134217728 |        1.00 |      16 |       16 |  0.054 |  0.110 |  0.290 |
| 128 |    67108864 |        0.50 |      16 |       16 |  0.059 |  0.065 |  0.349 |
| 128 |    67108864 |        0.75 |      16 |       16 |  0.064 |  0.067 |  0.334 |
| 128 |    67108864 |        1.00 |      16 |       16 |  0.042 |  0.088 |  0.331 |
```

**.:: Master branch ::.**
```
mlanger@1ad660b621f8:/scratch/proj/merlin-kv/build$ ./merlin_hashtable_benchmark

## Benchmark

* GPU: 1 x NVIDIA A10: 8.6
* Key Type = uint64_t
* Value Type = float32 * {dim}
* Key-Values per OP = 1048576
* ***Throughput Unit: Billion-KV/second***

### On pure HBM mode:

| dim |    capacity | load_factor | HBM(GB) | HMEM(GB) | insert |   find |  erase |
|----:|------------:|------------:|--------:|---------:|-------:|-------:|-------:|
|  16 |    67108864 |        0.50 |      16 |        0 |  0.411 |  0.646 |  0.275 |
|  16 |    67108864 |        0.75 |      16 |        0 |  0.374 |  0.563 |  0.307 |
|  16 |    67108864 |        1.00 |      16 |        0 |  0.104 |  0.334 |  0.277 |
|  64 |    67108864 |        0.50 |      16 |        0 |  0.328 |  0.293 |  0.309 |
|  64 |    67108864 |        0.75 |      16 |        0 |  0.298 |  0.291 |  0.308 |
|  64 |    67108864 |        1.00 |      16 |        0 |  0.097 |  0.255 |  0.271 |

### On HBM+HMEM hybrid mode:

| dim |    capacity | load_factor | HBM(GB) | HMEM(GB) | insert |   find |  erase |
|----:|------------:|------------:|--------:|---------:|-------:|-------:|-------:|
|  64 |   134217728 |        0.50 |      16 |       16 |  0.100 |  0.102 |  0.277 |
|  64 |   134217728 |        0.75 |      16 |       16 |  0.095 |  0.102 |  0.273 |
|  64 |   134217728 |        1.00 |      16 |       16 |  0.057 |  0.098 |  0.267 |
| 128 |    67108864 |        0.50 |      16 |       16 |  0.062 |  0.061 |  0.272 |
| 128 |    67108864 |        0.75 |      16 |       16 |  0.062 |  0.063 |  0.337 |
| 128 |    67108864 |        1.00 |      16 |       16 |  0.041 |  0.080 |  0.292 |
```