@bashimao 2022-12-08
We are looking at significantly increased erase performance and slightly increased insert/find throughputs. But mind you that the main benefit of this approach is that it eliminates most (if not all) memory fragmentation.